feat: 거대대금 비율 업데이트 로직 구현
