--- frontend/src/entities/stock/model/useStockStore.ts
@@ -1,27 +1,0 @@
-import { create } from 'zustand';
-import { devtools } from 'zustand/middleware';
-import { type RealSectorName } from '@/entities/stock/types/stock.types';
-
-interface StockState {
-  // UI 상태(State)만 관리
-  selectedSectorId: RealSectorName;
-  hoveredTickerId: string | null;
-  
-  // 액션(Actions)
-  actions: {
-    setSelectedSector: (id: RealSectorName) => void;
-    setHoveredTicker: (ticker: string | null) => void;
-  }
-}
-
-export const useStockStore = create<StockState>()(
-  devtools((set) => ({
-    selectedSectorId: '전체',
-    hoveredTickerId: null,
-
-    actions: {
-      setSelectedSector: (id) => set({ selectedSectorId: id }), 
-      setHoveredTicker: (ticker) => set({ hoveredTickerId: ticker }),
-    },
-  }))
-);
