+++ frontend/src/widgets/heatmap/StockTooltip.tsx
