+++ frontend/src/widgets/news-feed/NewsFeed.tsx
